requirements.txt
fastapi
uvicorn
ultralytics
pillow
python-multipart
streamlit
requests
